+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/jphes/hierarchy/HierarchyStore.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.jphes_hierarchy;
+package org.hisp.dhis.jphes.hierarchy;
 
 import org.hisp.dhis.common.GenericIdentifiableObjectStore;
