+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/jphes/hierarchy/DefaultHierarchyService.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.jphes_hierarchy;
+package org.hisp.dhis.jphes.hierarchy;
 
 import java.util.List;
