+++ dhis-2/dhis-services/dhis-service-core/src/main/resources/i18n_global.properties
@@ -80,7 +80,7 @@ dhis-web-maintenance-rbf=Maintenance RBF
 
 #-- Jphes modules ----------------------------------------------------------#
 
-jphes-web-hierarchy=Partner Hierarchy
+jphes-web-hierarchy=JPHES Hierarchy
 
 #-- Intro modules -------------------------------------------------------------#
