+++ dhis-2/dhis-web/jphes-web-apps/jphes-web-hierarchy/src/main/java/org/hisp/dhis/jphes/hierarchy/action/GetPartnerListAction.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.hierarchy.action;
+package org.hisp.dhis.jphes.hierarchy.action;
 
 import org.hisp.dhis.jphes.hierarchy.PartnerHierarchy;
 import org.hisp.dhis.jphes.hierarchy.HierarchyService;
