+++ dhis-2/dhis-web/jphes-web-apps/jphes-web-hierarchy/src/main/java/org/hisp/dhis/jphes/hierarchy/action/NoAction.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.hierarchy.action;
+package org.hisp.dhis.jphes.hierarchy.action;
 
 
 import com.opensymphony.xwork2.Action;
