+++ dhis-2/dhis-web/jphes-web-apps/jphes-web-hierarchy/src/main/resources/struts.xml
@@ -10,13 +10,13 @@
 
     <!-- JPHES Web APP -->
 
-    <action name="index" class="org.hisp.dhis.hierarchy.action.NoAction">
+    <action name="index" class="org.hisp.dhis.jphes.hierarchy.action.NoAction">
       <result name="success" type="velocity">/main.vm</result>
       <param name="page">/jphes-web-hierarchy/index.vm</param>
       <param name="menu">/jphes-web-hierarchy/menu.vm</param>
     </action>
 
-    <action name="allPartner" class="org.hisp.dhis.hierarchy.action.GetPartnerListAction">
+    <action name="allPartner" class="org.hisp.dhis.jphes.hierarchy.action.GetPartnerListAction">
       <result name="success" type="velocity">/main.vm</result>
       <param name="page">/jphes-web-hierarchy/allPartner.vm</param>
       <param name="menu">/jphes-web-hierarchy/menu.vm</param>
@@ -24,12 +24,12 @@
       <!--<param name="requiredAuthorities">F_JPHES_LIST_PARTNER</param>-->
     </action>
 
-    <action name="showAddPartnerForm" class="org.hisp.dhis.hierarchy.action.NoAction">
+    <action name="showAddPartnerForm" class="org.hisp.dhis.jphes.hierarchy.action.NoAction">
       <result name="success" type="velocity">/main.vm</result>
       <param name="page">/jphes-web-hierarchy/addPartnerForm.vm</param>
     </action>
 
-    <action name="addPartner" class="org.hisp.dhis.hierarchy.action.AddPartnerAction">
+    <action name="addPartner" class="org.hisp.dhis.jphes.hierarchy.action.AddPartnerAction">
       <result name="success" type="redirect">allPartner.action?currentPage=${keyCurrentPage}&amp;key=${keyCurrentKey}</result>
       <result name="error" type="redirect">showAddPartnerForm.action</result>
     </action>
