+++ dhis-2/dhis-web/jphes-web-apps/jphes-web-hierarchy/src/main/webapp/jphes-web-hierarchy/allPartner.vm
@@ -7,7 +7,7 @@
   var i18n_none = '$encoder.jsEscape( $i18n.getString( "none" ) , "'")';
 </script>
 
-<h3>Hierarchy #openHelp( "hierarchy" )</h3>
+<h3>Hierarchy #openHelp( "jphes_hierarchy" )</h3>
 
 <div id="contextMenu" class="contextMenu">
   <ul id="contextMenuItems" class="contextMenuItems">
