+++ dhis-2/dhis-web/jphes-web-apps/jphes-web-hierarchy/src/main/webapp/jphes-web-hierarchy/index.vm
@@ -1,5 +1,8 @@
-<h3>$i18n.getString( "partners" )&nbsp; #openHelp( "partners" )</h3>
+<h3>$i18n.getString( "jphes_hierarchy" )&nbsp; #openHelp( "jphes_hierarchy" )</h3>
 
 <ul class="introList">
-    #introListImgItem( "allPartner.action" "hierarchy" "hierarchy" )
+    #introListImgItem( "allPartner.action" "jphes_national" "jphes_level" )
+    #introListImgItem( "allPartner.action" "jphes_donor" "jphes_level" )
+    #introListImgItem( "allPartner.action" "jphes_agency" "jphes_level" )
+    #introListImgItem( "allPartner.action" "jphes_mechanism" "jphes_level" )
 </ul>
