+++ dhis-2/dhis-web/jphes-web-apps/jphes-web-hierarchy/src/main/webapp/jphes-web-hierarchy/menu.vm
@@ -1,4 +1,7 @@
-<h2>Hierarchy&nbsp;</h2>
+<h2>$i18n.getString( "jphes_hierarchy" )&nbsp;</h2>
 <ul>
-    <li><a href="allPartner.action">Hierarchy&nbsp;</a></li>
+  <li><a href="allPartner.action">$i18n.getString( "jphes_national" )&nbsp;</a></li>
+  <li><a href="allPartner.action">$i18n.getString( "jphes_donor" )&nbsp;</a></li>
+  <li><a href="allPartner.action">$i18n.getString( "jphes_agency" )&nbsp;</a></li>
+  <li><a href="allPartner.action">$i18n.getString( "jphes_mechanism" )&nbsp;</a></li>
 </ul>
